Refactor code, clean cli.js file. Create a folder with the name project
