+++ package.json
@@ -18,7 +18,7 @@
     "graphql",
     "architecture"
   ],
-  "author": "LuisPe <luispedrotoloy@gmail.com>",
+  "author": "LuisPe <https://github.com/LuisPe>",
   "repository": {
     "type": "git",
     "url": "https://github.com/LuisPe/inka.git"
